UAR-581: Refactor test files
