+++ src/test/java/uk/gov/companieshouse/overseasentitiesapi/model/PersonNameTest.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.overseasentitiesapi.service.changelist;
+package uk.gov.companieshouse.overseasentitiesapi.model;
 
 import static org.junit.jupiter.api.Assertions.*;
