+++ src/test/java/uk/gov/companieshouse/overseasentitiesapi/service/BeneficialOwnerCessationServiceTest.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.overseasentitiesapi.model.updatesubmission.service.cessations;
+package uk.gov.companieshouse.overseasentitiesapi.service;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.*;
